Updated: playlist.py
I added the add_song() method to handle what happens when the user tries to add nodes into an empty or full linked list, alongside the display_playlist() function to visually send back an output showing that the code worked. This is also connected to the playlist_display() function which returns each node's individual title and artist.